Разделил utils.js на render.js и common.js. Отредактировал импорты
