+++ src/components/Task-editor.js
@@ -1,5 +1,5 @@
 import {COLORS, DAYS, MONTH_NAMES} from "../constants.js";
-import {formatTime} from "../utils.js";
+import {formatTime} from "../utils/common.js";
 import AbstractComponent from "./abstract-component.js";
